Port DetectorManufacturer
