+++ bundle/edu.gemini.pot/src/main/java/edu/gemini/spModel/gemini/ghost/GhostType.java
@@ -36,57 +36,6 @@ public class GhostType {
 
     public static final double E2V_PIXEL_SIZE = 0.4; // arcsec/pixel
 
-    public enum DetectorManufacturer implements DisplayableSpType {
-        BLUE("E2V_CCD231-84-1-G57", "Blue", E2V_PIXEL_SIZE, 4096, 4112, 0.0003263888888888889),
-        RED ("E2V_CCD231-C6-1-G58", "Red", E2V_PIXEL_SIZE, 6144, 4096, 0.00022916666666666666);
-
-        public static final GhostType.DetectorManufacturer DEFAULT = RED;
-        public static final ItemKey KEY = new ItemKey(INSTRUMENT_KEY, "detectorManufacturer");
-
-        private final String _displayValue;
-        private final String _manufacter;
-        private final double _pixelSize;
-        private final int _xSize;
-        private final int _ySize;
-
-        private final double _darkCurrent;
-
-        DetectorManufacturer(final String manufacter, final String displayValue,
-                             final double pixelSize, final int xSize, final int ySize, final double darkCurrent) {
-            this._manufacter = manufacter;
-            this._displayValue = displayValue;
-            this._pixelSize = pixelSize;
-            this._xSize = xSize;
-            this._ySize = ySize;
-            this._darkCurrent = darkCurrent;
-        }
-
-        public String displayValue() {
-            return _displayValue;
-        }
-
-        /**
-         * arcsec/pixel
-         */
-        public double pixelSize() {
-            return _pixelSize;
-        }
-
-        public int getXsize() {
-            return _xSize;
-        }
-
-        public int getYsize() {
-            return _ySize;
-        }
-
-        public double get_darkCurrent() { return _darkCurrent;}
-
-        public String getManufacter() {
-            return _manufacter;
-        }
-
-    }
 
 //    public enum Resolution implements DisplayableSpType, LoggableSpType, SequenceableSpType {
 //        STANDARD("SR"),
